Move metadata-webhook from experimental to serving directory
