+++ Makefile
@@ -53,7 +53,7 @@ test-unit:
 	go test ./knative-operator/...
 	go test ./openshift-knative-operator/...
 	go test ./serving/ingress/...
-	go test ./experimental/metadata-webhook/...
+	go test ./serving/metadata-webhook/...
 
 # Run only SERVING/EVENTING E2E tests from the current repo.
 test-e2e:
